+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-07-15 19:42:46
-// Commit: test(core): add unit tests for core business logic
+// Updated at: 2026-07-17 18:26:46
+// Commit: perf(core): apply RepaintBoundary on animated core widgets
 
 class coreComponentState {
-  final String commitRef = '2871';
-  final String lastMessage = 'test(core): add unit tests for core business logic';
-  final DateTime timestamp = DateTime.parse('2026-07-15T19:42:46+05:00');
+  final String commitRef = '2921';
+  final String lastMessage = 'perf(core): apply RepaintBoundary on animated core widgets';
+  final DateTime timestamp = DateTime.parse('2026-07-17T18:26:46+05:00');
 }